refactor: use pacman to install packages

+++ bat/init.sh
@@ -4,8 +4,14 @@ echo '****************************************'
 echo '*            Installing bat            *'
 echo '****************************************'
 
-curl -L 'https://github.com/sharkdp/bat/releases/download/v0.24.0/bat-musl_0.24.0_amd64.deb' -o 'bat.deb'
+if [[ $INSTALL_OS = 'arch' ]]; then
+  yes | sudo pacman -S bat
+fi
 
-sudo dpkg -i bat.deb
+if [[ $INSTALL_OS = 'popos' ]]; then
+  curl -L 'https://github.com/sharkdp/bat/releases/download/v0.24.0/bat-musl_0.24.0_amd64.deb' -o 'bat.deb'
 
-rm bat.deb
+  sudo dpkg -i bat.deb
+
+  rm bat.deb
+fi
