+++ fira-code/init.sh
@@ -4,14 +4,20 @@ echo '********************************************'
 echo '*           Installing Fira Code           *'
 echo '********************************************'
 
-FONT_VERSION=$(curl -s https://api.github.com/repos/ryanoasis/nerd-fonts/releases/latest | grep -Po '"tag_name": "v\K[^"]*')
-curl -Lo fira_code.zip "https://github.com/ryanoasis/nerd-fonts/releases/latest/download/FiraCode.zip"
+if [[ $INSTALL_OS = 'arch' ]]; then
+  yes | sudo pacman -S ttf-firacode-nerd
+fi
 
-mkdir tmp
-unzip fira_code.zip -d tmp
+if [[ $INSTALL_OS = 'popos' ]]; then
+  FONT_VERSION=$(curl -s https://api.github.com/repos/ryanoasis/nerd-fonts/releases/latest | grep -Po '"tag_name": "v\K[^"]*')
+  curl -Lo fira_code.zip "https://github.com/ryanoasis/nerd-fonts/releases/latest/download/FiraCode.zip"
 
-mkdir -p "$HOME"/.local/share/fonts
-mv tmp/*.ttf "$HOME"/.local/share/fonts
+  mkdir tmp
+  unzip fira_code.zip -d tmp
 
-rm fira_code.zip
-rm -r tmp
+  mkdir -p "$HOME"/.local/share/fonts
+  mv tmp/*.ttf "$HOME"/.local/share/fonts
+
+  rm fira_code.zip
+  rm -r tmp
+fi
